jira TSI-62: implementation, fix bug, UT and fix UT failed

+++ cliniko-connect/src/main/java/com/tq/cliniko/lambda/handler/ClinikoDisconnectHandler.java
@@ -3,10 +3,12 @@ package com.tq.cliniko.lambda.handler;
 import com.tq.cliniko.lambda.exception.ClinikoConnectException;
 import com.tq.cliniko.lambda.model.ClinikoPractitionerConnectReq;
 import com.tq.cliniko.lambda.resp.ClinikoConnectStatusResponse;
+import com.tq.common.lambda.dynamodb.model.ClinikoCompanyInfo;
 import com.tq.common.lambda.dynamodb.model.ClinikoSbmSync;
 import com.tq.common.lambda.dynamodb.model.ClinikoSyncStatus;
 import com.tq.common.lambda.dynamodb.model.SbmBookingList;
 import com.tq.common.lambda.dynamodb.model.SbmSyncFutureBookings;
+import com.tq.common.lambda.dynamodb.service.ClinikoCompanyInfoService;
 import com.tq.common.lambda.dynamodb.service.ClinikoItemService;
 import com.tq.common.lambda.dynamodb.service.ClinikoSyncToSbmService;
 import com.tq.common.lambda.dynamodb.service.SbmListBookingService;
@@ -17,13 +19,15 @@ public class ClinikoDisconnectHandler implements ConnectHandler {
 	private ClinikoItemService clinikoItemService = null;
 	private SbmSyncFutureBookingsService sbmSyncFutureBookingService = null;
 	private SbmListBookingService sbmListBookingService = null;
+	private ClinikoCompanyInfoService clinikoCompanyService;
 
 	public ClinikoDisconnectHandler(ClinikoSyncToSbmService clinikoSyncService, ClinikoItemService clinikoItemService,
-			SbmSyncFutureBookingsService sbmSyncFutureBookingService, SbmListBookingService sbmListBookingService) {
+			SbmSyncFutureBookingsService sbmSyncFutureBookingService, SbmListBookingService sbmListBookingService, ClinikoCompanyInfoService clinikoCompanyService) {
 		this.clinikoSyncService = clinikoSyncService;
 		this.clinikoItemService = clinikoItemService;
 		this.sbmSyncFutureBookingService = sbmSyncFutureBookingService;
 		this.sbmListBookingService = sbmListBookingService;
+		this.clinikoCompanyService = clinikoCompanyService;
 	}
 
 	@Override
@@ -39,9 +43,11 @@ public class ClinikoDisconnectHandler implements ConnectHandler {
 				SbmSyncFutureBookings sbmSyncFutureBookings = sbmSyncFutureBookingService
 						.load(clinikoSbmSync.getSbmId());
 				SbmBookingList sbmListBooking = sbmListBookingService.load(clinikoSbmSync.getSbmId());
-				if (sbmSyncFutureBookings != null && sbmListBooking != null) {
+				ClinikoCompanyInfo clinikoCompanyInfo = clinikoCompanyService.load(clinikoSbmSync.getApiKey());
+				if (sbmSyncFutureBookings != null && sbmListBooking != null && clinikoCompanyInfo != null) {
 					sbmSyncFutureBookingService.delete(sbmSyncFutureBookings);
 					sbmListBookingService.delete(sbmListBooking);
+					clinikoCompanyService.delete(clinikoCompanyInfo);
 				}
 			}
 		} else {
